+++ packages/block-basic-image/src/svg-image.tsx
@@ -1,0 +1,13 @@
+import * as React from "react"
+
+export function SvgImage(props) {
+  return (
+    <svg width="1em" height="1em" viewBox="0 0 24 24" fill="none" {...props}>
+      <path
+        d="M4.828 21l-.02.02-.021-.02H2.992A.993.993 0 012 20.007V3.993A1 1 0 012.992 3h18.016c.548 0 .992.445.992.993v16.014a1 1 0 01-.992.993H4.828zM20 15V5H4v14L14 9l6 6zm0 2.828l-6-6L6.828 19H20v-1.172zM8 11a2 2 0 110-4 2 2 0 010 4z"
+        fill="currentColor"
+      />
+    </svg>
+  )
+}
+