+++ packages/block-code/src/svg.tsx
@@ -1,4 +1,4 @@
-import * as React from "react";
+import * as React from 'react';
 
 function SvgCode({ title, titleId, ...props }, svgRef) {
   return (
@@ -7,14 +7,10 @@ function SvgCode({ title, titleId, ...props }, svgRef) {
       height="1em"
       viewBox="0 0 24 24"
       fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-      ref={svgRef}
-      aria-labelledby={titleId}
       {...props}
     >
-      {title ? <title id={titleId}>{title}</title> : null}
       <path
-        d="M24 12l-5.657 5.657-1.414-1.414L21.172 12l-4.243-4.243 1.414-1.414L24 12zM2.828 12l4.243 4.243-1.414 1.414L0 12l5.657-5.657L7.07 7.757 2.828 12zm6.96 9H7.66l6.552-18h2.128L9.788 21z"
+        d="M8 21v-2h3V5H8V3h8v2h-3v14h3v2H8zM18.55 7.05L23.5 12l-4.95 4.95-1.414-1.414L20.672 12l-3.536-3.536L18.55 7.05zm-12.928 0l1.414 1.414L3.5 12l3.536 3.536-1.414 1.414L.672 12l4.95-4.95z"
         fill="currentColor"
       />
     </svg>