+++ packages/editor/src/components/icons/h1.tsx
@@ -1,0 +1,25 @@
+import * as React from "react";
+
+function SvgH1({ title, titleId, ...props }, svgRef) {
+  return (
+    <svg
+      width="1em"
+      height="1em"
+      viewBox="0 0 24 24"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+      ref={svgRef}
+      aria-labelledby={titleId}
+      {...props}
+    >
+      {title ? <title id={titleId}>{title}</title> : null}
+      <path
+        d="M13 20h-2v-7H4v7H2V4h2v7h7V4h2v16zm8-12v12h-2v-9.796l-2 .536V8.67L19.5 8H21z"
+        fill="currentColor"
+      />
+    </svg>
+  );
+}
+
+const ForwardRef = React.forwardRef(SvgH1);
+export default ForwardRef;