+++ packages/editor/src/utils/blocks.ts
@@ -18,7 +18,7 @@ interface BlockConfig<Data, Config> {
   defaultValue: Partial<Data>;
   isEmpty?: (d: Data) => boolean;
   mdast?: {
-    type: BlockContent['type'] | `${BlockContent['type']}.${PhrasingContent['type']}}`,
+    type: BlockContent['type'] | `${BlockContent['type']}.${PhrasingContent['type']}`,
     in?: (content: BlockContent | PhrasingContent | any, toHTML: (child: BlockContent['children']) => HTMLElement) => Data
   },
   tune?: Array<{