+++ packages/editor/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @hexx/editor
 
+## 0.6.10
+
+### Patch Changes
+
+- - Update Icon
+  - Add menu prop on plus button
+  - add config on editor focus method
+
 ## 0.6.9
 
 ### Patch Changes